chore: move observatory
