+++ api/src/pdc/services/observatory/ObservatoryServiceProvider.ts
@@ -13,23 +13,23 @@ import { CampaignsAction } from "@/pdc/services/observatory/actions/incentiveCam
 import { AiresCovoiturageAction } from "@/pdc/services/observatory/actions/infra/AiresCovoiturageAction.ts";
 import { KeyfiguresAction } from "@/pdc/services/observatory/actions/keyfigures/KeyfiguresAction.ts";
 import { LocationAction } from "@/pdc/services/observatory/actions/location/LocationAction.ts";
-import { binding as JourneysByDistancesBinding } from "@/shared/observatory/distribution/journeysByDistances.schema.ts";
-import { binding as JourneysByHoursBinding } from "@/shared/observatory/distribution/journeysByHours.schema.ts";
-import { binding as GetBestFluxBinding } from "@/shared/observatory/flux/getBestFlux.schema.ts";
-import { binding as GetEvolFluxBinding } from "@/shared/observatory/flux/getEvolFlux.schema.ts";
-import { binding as GetFluxBinding } from "@/shared/observatory/flux/getFlux.schema.ts";
-import { binding as GetIncentiveBinding } from "@/shared/observatory/incentive/getIncentive.schema.ts";
-import { binding as CampaignsBinding } from "@/shared/observatory/incentiveCampaigns/campaigns.schema.ts";
-import { binding as AiresCovoiturageBinding } from "@/shared/observatory/infra/airesCovoiturage.schema.ts";
-import { binding as GetKeyfiguresBinding } from "@/shared/observatory/keyfigures/getKeyfigures.schema.ts";
-import { binding as LocationBinding } from "@/shared/observatory/location/location.schema.ts";
-import { binding as GetBestTerritoriesBinding } from "../../../shared/observatory/occupation/getBestTerritories.schema.ts";
-import { binding as GetEvolOccupationBinding } from "../../../shared/observatory/occupation/getEvolOccupation.schema.ts";
-import { binding as GetOccupationBinding } from "../../../shared/observatory/occupation/getOccupation.schema.ts";
 import { BestTerritoriesAction } from "./actions/occupation/BestTerritoriesAction.ts";
 import { EvolOccupationAction } from "./actions/occupation/EvolOccupationAction.ts";
 import { OccupationAction } from "./actions/occupation/OccupationAction.ts";
 import { config } from "./config/index.ts";
+import { binding as JourneysByDistancesBinding } from "./contracts/distribution/journeysByDistances.schema.ts";
+import { binding as JourneysByHoursBinding } from "./contracts/distribution/journeysByHours.schema.ts";
+import { binding as GetBestFluxBinding } from "./contracts/flux/getBestFlux.schema.ts";
+import { binding as GetEvolFluxBinding } from "./contracts/flux/getEvolFlux.schema.ts";
+import { binding as GetFluxBinding } from "./contracts/flux/getFlux.schema.ts";
+import { binding as GetIncentiveBinding } from "./contracts/incentive/getIncentive.schema.ts";
+import { binding as CampaignsBinding } from "./contracts/incentiveCampaigns/campaigns.schema.ts";
+import { binding as AiresCovoiturageBinding } from "./contracts/infra/airesCovoiturage.schema.ts";
+import { binding as GetKeyfiguresBinding } from "./contracts/keyfigures/getKeyfigures.schema.ts";
+import { binding as LocationBinding } from "./contracts/location/location.schema.ts";
+import { binding as GetBestTerritoriesBinding } from "./contracts/occupation/getBestTerritories.schema.ts";
+import { binding as GetEvolOccupationBinding } from "./contracts/occupation/getEvolOccupation.schema.ts";
+import { binding as GetOccupationBinding } from "./contracts/occupation/getOccupation.schema.ts";
 import { DistributionRepositoryProvider } from "./providers/DistributionRepositoryProvider.ts";
 import { FluxRepositoryProvider } from "./providers/FluxRepositoryProvider.ts";
 import { IncentiveCampaignsRepositoryProvider } from "./providers/IncentiveCampaignsRepositoryProvider.ts";
