+++ api/src/pdc/services/observatory/actions/incentiveCampaigns/CampaignsAction.ts
@@ -1,13 +1,13 @@
-import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { handler } from "@/ilos/common/index.ts";
+import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
 
-import { alias } from "@/shared/observatory/incentiveCampaigns/campaigns.schema.ts";
 import {
   handlerConfig,
   ParamsInterface,
   ResultInterface,
-} from "@/shared/observatory/incentiveCampaigns/campaigns.contract.ts";
+} from "../../contracts/incentiveCampaigns/campaigns.contract.ts";
+import { alias } from "../../contracts/incentiveCampaigns/campaigns.schema.ts";
 import { IncentiveCampaignsRepositoryInterfaceResolver } from "../../interfaces/IncentiveCampaignsRepositoryProviderInterface.ts";
 
 @handler({
