+++ api/src/pdc/services/observatory/actions/infra/AiresCovoiturageAction.ts
@@ -1,13 +1,9 @@
-import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { handler } from "@/ilos/common/index.ts";
+import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
 
-import { alias } from "@/shared/observatory/infra/airesCovoiturage.schema.ts";
-import {
-  handlerConfig,
-  ParamsInterface,
-  ResultInterface,
-} from "@/shared/observatory/infra/airesCovoiturage.contract.ts";
+import { handlerConfig, ParamsInterface, ResultInterface } from "../../contracts/infra/airesCovoiturage.contract.ts";
+import { alias } from "../../contracts/infra/airesCovoiturage.schema.ts";
 import { InfraRepositoryInterfaceResolver } from "../../interfaces/InfraRepositoryProviderInterface.ts";
 
 @handler({
