+++ api/src/pdc/services/observatory/contracts/distribution/journeysByHours.schema.ts
@@ -1,10 +1,10 @@
-import { perimeterTypes } from "../../geo/shared/Perimeter.ts";
+import { directionTypes, perimeterTypes } from "../../../../../shared/geo/shared/Perimeter.ts";
 
-export const alias = "observatory.getLocation";
+export const alias = "observatory.journeysByHours";
 export const schema = {
   type: "object",
   additionalProperties: false,
-  required: ["year", "zoom", "type", "code"],
+  required: ["year", "type", "code"],
   properties: {
     year: {
       type: "integer",
@@ -25,11 +25,6 @@ export const schema = {
       minimum: 1,
       maximum: 2,
     },
-    zoom: {
-      type: "integer",
-      minimum: 0,
-      maximum: 8,
-    },
     type: {
       type: "string",
       enum: perimeterTypes,
@@ -41,6 +36,10 @@ export const schema = {
         { macro: "country" },
         { macro: "siren" },
       ],
+    },
+    direction: {
+      type: "string",
+      enum: directionTypes,
     },
   },
 };
