+++ api/src/pdc/services/observatory/contracts/flux/getBestFlux.schema.ts
