+++ api/src/pdc/services/observatory/contracts/flux/getFlux.contract.ts
