+++ api/src/pdc/services/observatory/contracts/incentive/getIncentive.schema.ts
