+++ api/src/pdc/services/observatory/contracts/infra/airesCovoiturage.schema.ts
@@ -1,0 +1,19 @@
+import { perimeterTypes } from "../../../../../shared/geo/shared/Perimeter.ts";
+
+export const alias = "observatory.airesCovoiturage";
+export const schema = {
+  type: "object",
+  additionalProperties: false,
+  required: ["type", "code"],
+  properties: {
+    type: {
+      type: "string",
+      enum: perimeterTypes,
+    },
+    code: {
+      anyOf: [{ macro: "insee" }, { macro: "department" }, { macro: "country" }, { macro: "siren" }],
+    },
+  },
+};
+
+export const binding = [alias, schema];
