+++ api/src/pdc/services/observatory/contracts/occupation/getBestTerritories.contract.ts
