+++ api/src/pdc/services/observatory/contracts/occupation/getOccupation.schema.ts
