+++ api/src/pdc/services/observatory/interfaces/LocationRepositoryProviderInterface.ts
@@ -2,13 +2,9 @@ import type {
   ParamsInterface as LocationParamsInterface,
   ResultInterface as LocationResultInterface,
   SqlResultInterface as LocationSqlResultInterface,
-} from "@/shared/observatory/location/location.contract.ts";
+} from "../contracts/location/location.contract.ts";
 
-export type {
-  LocationParamsInterface,
-  LocationResultInterface,
-  LocationSqlResultInterface,
-};
+export type { LocationParamsInterface, LocationResultInterface, LocationSqlResultInterface };
 
 export interface LocationRepositoryInterface {
   getLocation(
@@ -16,8 +12,7 @@ export interface LocationRepositoryInterface {
   ): Promise<LocationResultInterface>;
 }
 
-export abstract class LocationRepositoryInterfaceResolver
-  implements LocationRepositoryInterface {
+export abstract class LocationRepositoryInterfaceResolver implements LocationRepositoryInterface {
   async getLocation(
     params: LocationParamsInterface,
   ): Promise<LocationResultInterface> {
